risk-analysis/requirements.txt
blinker
click
Flask
itsdangerous
Jinja2
MarkupSafe
numpy
pandas
pip
psycopg2-binary
PyMuPDF
python-dateutil
python-dotenv
pytz
six
tzdata
Werkzeug
requests